refactor register component
