+++ frontend/AngularApp/src/app/register/register.component.scss
@@ -1,0 +1,41 @@
+@import "../../styles.scss";
+
+.tooltip-icon {
+  cursor: pointer;
+  display: inline-block;
+  margin-left: 0.5rem;
+  background-color: $pink;
+  color: $white;
+  border-radius: 50%;
+  width: 1.1rem;
+  height: 1.1rem;
+  text-align: center;
+  line-height: 1.1rem;
+  font-size: 1rem;
+  position: relative;
+
+  &:hover + .tooltip-content {
+    display: block;
+  }
+}
+
+.tooltip-content {
+  display: none;
+  position: absolute;
+  background-color: $white;
+  border: 1px solid $black;
+  padding: 0.5rem;
+  border-radius: 0.25rem;
+  box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.1);
+  width: 200px;
+  top: 80%;
+  left: 48%;
+  z-index: 1000;
+  font-size: 1rem;
+
+  ul {
+    margin: 0;
+    padding-left: 1rem;
+    font-size: 0.8rem;
+  }
+}
